Trying to fix Output by changing malloc

+++ 0x0C-more_malloc_free/100-realloc.c
@@ -38,7 +38,7 @@ void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 		return (NULL);
 	}
 
-	c_s = (old_size < new_size) ? old_size : new_size;
+	c_s = (size_t)((old_size < new_size) ? old_size : new_size);
 	memcpy(n_p, ptr, c_s);
 
 	free(ptr);
